refactor: Renamed NewMessageModal to MessageModal and moved inside components/ui
