+++ resources/js/pages/Index.vue
@@ -1,13 +1,13 @@
 <script setup>
-import Header from "../components/layouts/Header.vue";
-import Sidebar from "../components/layouts/Sidebar.vue";
-import Container from "../components/layouts/Container.vue";
-import Channels from "../components/Channels.vue";
-import Contacts from "../components/Contacts.vue";
-import NewMessageModal from "../components/NewMessageModal.vue"
-import Content from "../components/layouts/Content.vue";
+import Header from "@/components/layouts/Header.vue";
+import Sidebar from "@/components/layouts/Sidebar.vue";
+import Container from "@/components/layouts/Container.vue";
+import Channels from "@/components/layouts/Channels.vue";
+import Contacts from "@/components/layouts/Contacts.vue";
+import MessageModal from "@/components/MessageModal.vue"
+import Content from "@/components/layouts/Content.vue";
 
-import { useGlobalStore } from "../stores/global";
+import { useGlobalStore } from "@/stores/global";
 
 const props = defineProps({
   contacts: Array,
@@ -26,7 +26,7 @@ const store = useGlobalStore();
         <Contacts :contacts="contacts" />
       </Sidebar>
       <Content />
-      <NewMessageModal v-if="store.newMessageModal" :channels="channels" @close="store.toggleNewMessageModal(false)"/>
+      <MessageModal v-if="store.newMessageModal" :channels="channels" @close="store.toggleNewMessageModal(false)"/>
     </Container>
   </div>
 </template>
